Refactors code to handle commands for delete and complete task

+++ commands/complete.js
@@ -1,5 +1,8 @@
 const db = require('../dbqueries');
 
-module.exports = (taskID) => {
-  db.deleteTask(taskID)
+module.exports = (command, taskID) => {
+  if(command == 'complete')
+    db.completeTask(taskID)
+  else
+    db.deleteTask(taskID)
 }
